Added cache option
This feature ended up requiring a lot of the plugin to be rewritten (for optimization). I also took the opportunity to rename some option names. I created a mapping implementation that will warn users of deprecated option names.

Also, I will likely be updating the entire plugin soon so it is more prototype based. It will help make it more clear as to where functionality lives.